Crate umio

Message Based Readiness API

This library is a thin wrapper around mio for clients who wish to use a single udp socket in conjunction with message passing and timeouts.

Modules

external
Exports of bare mio types.

Structs

ELoop
Wrapper around the main application event loop.

ELoopBuilder
Builder for specifying attributes of an event loop.

Provider
Provides services to dispatcher clients.

Traits

Dispatcher
Handles events occurring within the event loop.