Struct umio::ELoopBuilder

pub struct ELoopBuilder { /* fields omitted */ }

Builder for specifying attributes of an event loop.

Methods

impl ELoopBuilder

fn new() -> ELoopBuilder

Create a new event loop builder.

fn channel_capacity(self, capacity: usize) -> ELoopBuilder

Manually set the maximum channel message capacity.

fn timer_capacity(self, capacity: usize) -> ELoopBuilder

Manually set the maximum timer capacity.

fn bind_address(self, address: SocketAddr) -> ELoopBuilder

Manually set the bind address for the udp socket in the event loop.

fn buffer_length(self, length: usize) -> ELoopBuilder

Manually set the length of buffers provided by the event loop.

fn build<D: Dispatcher>(self) -> Result<ELoop<D>>

Build the event loop with the current builder.