Struct umio::ELoop

pub struct ELoop<D: Dispatcher> { /* fields omitted */ }

Wrapper around the main application event loop.

Methods

impl<D: Dispatcher> ELoop<D>

fn channel(&self) -> Sender<D::Message>

Grab a channel to send messages to the event loop.

fn run(&mut self, dispatcher: D) -> Result<()>

Run the event loop with the given dispatcher until a shutdown occurs.